+++ statsd/aggregator.go
@@ -47,7 +47,7 @@ func NewMetricAggregator(senders []backend.MetricSender, percentThresholds []flo
 	a.FlushInterval = flushInterval
 	a.ExpiryInterval = expiryInterval
 	a.Senders = senders
-	a.MetricQueue = make(chan types.Metric, maxQueueSize)
+	a.MetricQueue = make(chan types.Metric, maxQueueSize * 10) // we are going to receive more metrics than messages
 	a.MaxWorkers = maxWorkers
 	a.PercentThresholds = percentThresholds
 	a.Counters = types.Counters{}